Silence warning about certificate
The certificate in our current identity environment lacks a
subjectAltName, causing a fallback to check the commonName. This is soon
to be unsupported, thus raising a SecurityWarning. As this is purely a
warning and there is no current security issue, we are silencing it.
Note that the warning is only silenced when coming from the urllib3
that is included with requests, meaning we're not blanket silencing
anything.